requirements.txt
anthropic>=0.84.0
APScheduler>=3.11.0
python-dotenv>=1.0.0
python-telegram-bot>=22.0
Telethon>=1.42.0
